* src/hc08/gen.c: fixed MSVC warning C4146: unary minus operator applied to   unsigned type, result still unsigned
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@4922 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/hc08/gen.c
@@ -3412,7 +3412,7 @@ genMinusDec (iCode * ic)
           needpulh = FALSE;
         }
       loadRegFromAop (hc08_reg_hx, AOP(left), 0);
-      emitcode ("aix","#%d", -icount);
+      emitcode ("aix","#%d", -(int) icount);
       hc08_dirtyReg (hc08_reg_hx, FALSE);
       storeRegToAop (hc08_reg_hx, AOP(result), 0);
       pullOrFreeReg (hc08_reg_h, needpulh);
